# schemas/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "X2KNWLDG local API — v1",
    "version": "1.0.0",
    "summary": "The frozen read-only HTTP contract of the Knowledge Canvas local web layer.",
...
    "license": {
      "name": "MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "description": "Loopback only. The server must not bind a non-loopback interface (T-108). The port is an operational default that T-116 settles; it is not part of the frozen contract.",
      "variables": {
        "port": {
          "default": "8765"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/sources/{source_id}/entities": {
      "get": {
        "operationId": "listSourceEntities",
        "tags": [
          "sources",
          "entities"
        ],
        "summary": "The knowledge units of one source",
        "description": "One page of `EntityRef` records belonging to *source_id*. In v1 these are knowledge units only: captions, segments, and coverage windows are reserved in the vocabulary and deliberately unemitted (D-023).",
        "parameters": [
          {
            "$ref": "#/components/parameters/SourceIdPath"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/ProvenanceClassQuery"
          },
          {
            "$ref": "#/components/parameters/KindQuery"
          },
          {
            "$ref": "#/components/parameters/MinConfidenceQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of entities.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/IndexUnavailable"
          }
        }
      }
    },
...
    "/api/entities/{entity_id}": {
      "get": {
        "operationId": "getEntity",
        "tags": [
          "entities"
        ],
        "summary": "One entity by global id",
        "description": "An `EntityRef` addressed by its three-part global id (D-011). A canonical concept is addressed as `library:concepts:<hash>` (D-016).",
        "parameters": [
          {
            "$ref": "#/components/parameters/EntityIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "The entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/IndexUnavailable"
          }
        }
      }
    },
    "/api/artifacts/{artifact_id}": {
      "get": {
        "operationId": "getArtifact",
        "tags": [
          "artifacts"
        ],
        "summary": "One artifact record",
        "description": "The `Artifact` record — what the file is, where it sits, whether it was present at index time. It is metadata about the file, never the file. Bytes come from `/api/media/{artifact_id}`. An artifact that was absent at index time is returned with `available: false` and a `200`: its record exists and the honest answer is that the file does not.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArtifactIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "The artifact record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/IndexUnavailable"
          }
        }
      }
    },
...
    "/api/graph": {
      "get": {
        "operationId": "getGraph",
        "tags": [
          "graph"
        ],
        "summary": "Nodes and edges for the Knowledge Map",
        "description": "`EntityRef` nodes and `IndexedRelation` edges, bounded by `limit`. Every edge states its `relation_vocabulary`, so the Map can style the two library-synthetic edge types — `derived_from` and `expresses_concept`, the most common edges in the current data — without presenting them as canonical relations.\n\n`truncated` is stated rather than implied: a Map that has silently received part of a graph would draw a false picture of what is connected.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/SourceIdQuery"
          },
          {
            "$ref": "#/components/parameters/ProvenanceClassQuery"
          },
          {
            "$ref": "#/components/parameters/RelationVocabularyQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "A bounded slice of the graph.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/IndexUnavailable"
          }
        }
      }
    },
    "/api/graph/neighborhood/{entity_id}": {
      "get": {
        "operationId": "getNeighborhood",
        "tags": [
          "graph"
        ],
        "summary": "The neighborhood of one entity",
        "description": "Nodes and edges within `depth` hops of *entity_id*, for the Map's incremental loading. The centre is echoed back so a client that batches requests cannot mis-attribute a response.",
        "parameters": [
          {
            "$ref": "#/components/parameters/EntityIdPath"
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "description": "Hops from the centre.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 3,
              "default": 1
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/RelationVocabularyQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "The neighborhood.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborhoodResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/IndexUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum records in the page.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 50
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque continuation token from a previous `page.next_cursor`. Opaque on purpose: the encoding belongs to the index (T-007), and a client that parses it has coupled itself to a private detail. A cursor the server cannot read is `400 invalid_request`.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
      "SourceIdPath": {
        "name": "source_id",
        "in": "path",
        "required": true,
        "description": "Two-part source id, `<source_type>:<external_id>`. A colon is a legal path character (RFC 3986 `pchar`) and needs no escaping. Resolved through `pipeline.resolve_run_dir`, which rejects rather than sanitises (D-020).",
        "schema": {
          "$ref": "../../v1/common.schema.json#/$defs/sourceId"
        }
      },
      "EntityIdPath": {
        "name": "entity_id",
        "in": "path",
        "required": true,
        "description": "Three-part global id, `<source_type>:<external_id>:<local_id>`.",
        "schema": {
          "$ref": "../../v1/common.schema.json#/$defs/globalId"
        }
      },
      "ArtifactIdPath": {
        "name": "artifact_id",
        "in": "path",
        "required": true,
        "description": "Three-part global id of an artifact, e.g. `youtube:pqlWNihgdjI:transcript`.",
        "schema": {
          "$ref": "../../v1/common.schema.json#/$defs/globalId"
        }
      },
      "SourceIdQuery": {
        "name": "source_id",
        "in": "query",
        "required": false,
        "description": "Restrict to one source.",
        "schema": {
          "$ref": "../../v1/common.schema.json#/$defs/sourceId"
        }
      },
...
      "ProvenanceClassQuery": {
        "name": "provenance_class",
        "in": "query",
        "required": false,
        "description": "Filter on origin: source evidence, recorded synthesis, or user content.",
        "schema": {
          "$ref": "../../v1/common.schema.json#/$defs/provenanceClass"
        }
      },
      "KindQuery": {
        "name": "kind",
        "in": "query",
        "required": false,
        "description": "Filter on knowledge kind.",
        "schema": {
          "$ref": "../../v1/common.schema.json#/$defs/knowledgeKind"
        }
      },
      "MinConfidenceQuery": {
        "name": "min_confidence",
        "in": "query",
        "required": false,
        "description": "Keep only entities whose copied confidence is at least this value. An entity with a null confidence is excluded, because no canonical file states a number for it and a default would be invented.",
        "schema": {
          "$ref": "../../v1/common.schema.json#/$defs/confidence"
        }
      },
      "RelationVocabularyQuery": {
        "name": "relation_vocabulary",
        "in": "query",
        "required": false,
        "description": "Filter on which of the three edge vocabularies an edge belongs to.",
        "schema": {
          "type": "string",
          "enum": [
            "canonical",
            "library_synthetic",
            "user"
          ]
        }
      }
    },
    "schemas": {
      "ApiVersion": {
        "description": "The API surface version. Fixed for the life of `schemas/api/v1/`. A breaking change becomes `schemas/api/v2/` and a `/api/v2/` path prefix, leaving these paths answering v1 (D-026).",
        "const": "v1"
      },
      "PageInfo": {
        "description": "Where a page sits in a collection.",
        "type": "object",
        "required": [
          "limit",
          "next_cursor"
        ],
        "additionalProperties": false,
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 500
          },
          "next_cursor": {
            "description": "Token for the next page, or null at the end of the collection.",
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 512
          },
          "total": {
            "description": "Total matching records, or null when the server did not count them. Null means unknown — it never means zero.",
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        }
      },
      "ErrorCode": {
        "description": "Closed error vocabulary. `invalid_id` is D-020 surfaced over HTTP: an id that fails `ids.py` or `resolve_run_dir` is refused before anything is read, and is reported as malformed rather than as absent. `unavailable` is the honest answer for an artifact whose record exists but whose file does not — the alternative, a placeholder, is forbidden by canvas plan §15.",
        "type": "string",
        "enum": [
          "invalid_id",
          "invalid_request",
          "not_found",
          "unavailable",
          "index_unavailable",
          "internal"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "description": "Human-readable, and safe to show. It names what was wrong with the request, never a host path.",
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          },
          "detail": {
            "description": "Structured context, or null. Never carries an absolute host path (risk R15).",
            "type": [
              "object",
              "null"
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "api_version",
          "schema_version",
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "$ref": "#/components/schemas/ApiVersion"
          },
          "schema_version": {
            "$ref": "../../v1/common.schema.json#/$defs/schemaVersion"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "ArtifactResponse": {
        "type": "object",
        "required": [
          "api_version",
          "schema_version",
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "$ref": "#/components/schemas/ApiVersion"
          },
          "schema_version": {
            "$ref": "../../v1/common.schema.json#/$defs/schemaVersion"
          },
          "data": {
            "$ref": "../../v1/artifact.schema.json"
          }
        }
      },
      "EntityResponse": {
        "type": "object",
        "required": [
          "api_version",
          "schema_version",
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "$ref": "#/components/schemas/ApiVersion"
          },
          "schema_version": {
            "$ref": "../../v1/common.schema.json#/$defs/schemaVersion"
          },
          "data": {
            "$ref": "../../v1/entity_ref.schema.json"
          }
        }
      },
      "EntityListResponse": {
        "type": "object",
        "required": [
          "api_version",
          "schema_version",
          "data",
          "page"
        ],
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "$ref": "#/components/schemas/ApiVersion"
          },
          "schema_version": {
            "$ref": "../../v1/common.schema.json#/$defs/schemaVersion"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "../../v1/entity_ref.schema.json"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
...
      "GraphPayload": {
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "truncated"
        ],
        "additionalProperties": false,
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "../../v1/entity_ref.schema.json"
            }
          },
          "edges": {
            "description": "The relations among the nodes on this page, at most `maxItems` of them. `limit` bounds the *nodes*, and an edge list is quadratic in the nodes it connects, so the two are not the same size; `truncated` is true when either was cut.",
            "type": "array",
            "items": {
              "$ref": "../../v1/indexed_relation.schema.json"
            },
            "maxItems": 5000
          },
          "truncated": {
            "description": "True when `limit` cut the result short. Stated rather than implied, so the Map never presents a partial graph as the whole one.",
            "type": "boolean"
          }
        }
      },
      "GraphResponse": {
        "type": "object",
        "required": [
          "api_version",
          "schema_version",
          "data",
          "page"
        ],
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "$ref": "#/components/schemas/ApiVersion"
          },
          "schema_version": {
            "$ref": "../../v1/common.schema.json#/$defs/schemaVersion"
          },
          "data": {
            "$ref": "#/components/schemas/GraphPayload"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
      "NeighborhoodPayload": {
        "type": "object",
        "required": [
          "center_id",
          "depth",
          "nodes",
          "edges",
          "truncated"
        ],
        "additionalProperties": false,
        "properties": {
          "center_id": {
            "$ref": "../../v1/common.schema.json#/$defs/globalId"
          },
          "depth": {
            "type": "integer",
            "minimum": 1,
            "maximum": 3
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "../../v1/entity_ref.schema.json"
            }
          },
          "edges": {
            "description": "The relations among the nodes on this page, at most `maxItems` of them. `limit` bounds the *nodes*, and an edge list is quadratic in the nodes it connects, so the two are not the same size; `truncated` is true when either was cut.",
            "type": "array",
            "items": {
              "$ref": "../../v1/indexed_relation.schema.json"
            },
            "maxItems": 5000
          },
          "truncated": {
            "type": "boolean"
          }
        }
      },
      "NeighborhoodResponse": {
        "type": "object",
        "required": [
          "api_version",
          "schema_version",
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "$ref": "#/components/schemas/ApiVersion"
          },
          "schema_version": {
            "$ref": "../../v1/common.schema.json#/$defs/schemaVersion"
          },
          "data": {
            "$ref": "#/components/schemas/NeighborhoodPayload"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request is malformed: a rejected id (`invalid_id`), or a bad parameter or unreadable cursor (`invalid_request`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "IndexUnavailable": {
        "description": "The index is absent, building, or in error, so no honest answer can be given yet (`index_unavailable`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "The server could not complete the request (`internal`). Declared because the handler that produces it exists: any unforeseen failure is rendered in this envelope rather than as a framework traceback, and a contract that does not say so is a contract a client cannot hold the server to.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "The path exists and does not answer this method (`invalid_request`). v1 is read-only, so every method but `GET` and `HEAD` is refused here; the response carries `Allow` naming what the path does accept, as RFC 9110 requires. Declared for the same reason `500` is: the router produces it, and a contract that does not say so is a contract a client cannot hold the server to.",
        "headers": {
          "Allow": {
            "description": "The methods this path answers.",
            "schema": {
              "type": "string"
            },
            "example": "GET, HEAD"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The id is well-formed but names nothing (`not_found`), or names an artifact whose file is absent (`unavailable`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}